requirements.txt
cachetools==6.2.2
dnslib==0.9.26
Flask==3.1.2
scapy==2.6.1
numpy==1.26.2
psutil==7.1.3
PyYAML==6.0.2
python-dotenv==1.2.1
torch==2.7.0
requests==2.31.0
watchdog==6.0.0
proxy.py==2.4.10
